add wx open sdk 1.8.0 and add choose invoice from weixin
