+++ plugin.xml
@@ -45,7 +45,7 @@
                 <true/>
             </dict>
         </config-file>
-        
+
 
         <config-file target="*-Info.plist" parent="CFBundleURLTypes">
             <array>
@@ -65,9 +65,9 @@
         <source-file src="src/ios/CDVWechat.m" />
 
         <!-- Wechat Official -->
-        <header-file src="src/ios/libs/OpenSDK1.7.5/WXApi.h" />
-        <header-file src="src/ios/libs/OpenSDK1.7.5/WXApiObject.h" />
-        <source-file src="src/ios/libs/OpenSDK1.7.5/libWeChatSDK.a" framework="true" />
+        <header-file src="src/ios/libs/OpenSDK1.8.0/WXApi.h" />
+        <header-file src="src/ios/libs/OpenSDK1.8.0/WXApiObject.h" />
+        <source-file src="src/ios/libs/OpenSDK1.8.0/libWeChatSDK.a" framework="true" />
 
         <!-- Other required frameworks -->
         <framework src="libz.tbd" />
@@ -92,7 +92,7 @@
             </feature>
             <preference name="WECHATAPPID" value="$WECHATAPPID"/>
         </config-file>
-        
+
         <config-file target="AndroidManifest.xml" parent="/*">
             <uses-permission android:name="android.permission.INTERNET"/>
             <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
